Hide scrollbars by fixing the textarea size
The width & height equal to 100% means that the textarea is 100% the
viewport plus twice the padding. That means the textarea is bigger than
the viewport.

Fix this by setting the textarea size to the viewport minus twice the
padding on both axis; that means the textarea is exactly the viewport
size.